Add profile panels. Add activities for patient information, journal entries and medical notes.
