+++ Connected_Health_App/app/src/main/res/layout/activity_profile_panels.xml
@@ -1,21 +1,26 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="fill_parent" android:layout_height="fill_parent"
-    android:orientation="horizontal">
+    xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
+    android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingRight="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    android:weightSum="3"
+    tools:context="au.edu.adelaide.connected_health_app.ProfilePanelsActivity">
 
-    <!--Info Panel-->
+    <!--Patient Information Panel-->
     <RelativeLayout
         android:layout_width="0dip"
         android:layout_height="fill_parent"
-        android:layout_weight="0.25"
-        android:onClick="goToInformationView"
-        android:background="@android:color/holo_orange_dark">
+        android:layout_weight="1"
+        android:onClick="goToProfileView"
+        android:background="@android:color/holo_red_light">
 
         <!--Panel label-->
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:textAppearance="?android:attr/textAppearanceLarge"
-            android:text="Information \nFor knowledge"
+            android:text="My Information"
             android:layout_alignParentTop="true"
             android:layout_centerHorizontal="true"
             android:layout_marginTop="46dp"
@@ -25,20 +30,20 @@
             android:textColor="#ffffffff" />
     </RelativeLayout>
 
-    <!--Questionnaire panel-->
+    <!--Journal Entries panel-->
     <RelativeLayout
         android:layout_width="0dip"
         android:layout_height="fill_parent"
-        android:layout_weight="0.25"
-        android:onClick="goToQuestionnaireView"
-        android:background="@android:color/holo_blue_dark">
+        android:layout_weight="1"
+        android:onClick="goToJournalView"
+        android:background="@android:color/holo_red_dark">
 
         <!--Panel label-->
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:textAppearance="?android:attr/textAppearanceLarge"
-            android:text="Questionnaires \nDaily activities"
+            android:text="My Journal"
             android:layout_alignParentTop="true"
             android:layout_centerHorizontal="true"
             android:layout_marginTop="46dp"
@@ -46,12 +51,12 @@
             android:textColor="#ffffffff" />
     </RelativeLayout>
 
-    <!--Profile panel-->
+    <!--Medical Notes panel-->
     <RelativeLayout
         android:layout_width="0dip"
         android:layout_height="fill_parent"
-        android:layout_weight="0.25"
-        android:onClick="goToProfilePanels"
+        android:layout_weight="1"
+        android:onClick="goToMedicalNotesView"
         android:background="@android:color/holo_red_light">
 
         <!--Panel label-->
@@ -59,28 +64,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:textAppearance="?android:attr/textAppearanceLarge"
-            android:text="Profile \nManage profile"
-            android:layout_alignParentTop="true"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="46dp"
-            android:gravity="center_horizontal"
-            android:textColor="#ffffffff" />
-    </RelativeLayout>
-
-    <!--Measurements panel-->
-    <RelativeLayout
-        android:layout_width="0dip"
-        android:layout_height="fill_parent"
-        android:layout_weight="0.25"
-        android:onClick="goToMeasurementView"
-        android:background="@android:color/holo_green_light">
-
-        <!--Panel label-->
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceLarge"
-            android:text="Measurements \nTo manage"
+            android:text="My Medical Notes"
             android:layout_alignParentTop="true"
             android:layout_centerHorizontal="true"
             android:layout_marginTop="46dp"
